+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Added
 * Support quickly change query and note_field in the pop up window. This will not change the default configuration.
 * Add "Quick AI" button to the browse window to quick click when checking the browse query, the query will be current input query in the browse window.
 * Change configuration files. Now it supports adding customized placeholders in prompts.
+* Change the response window. Now the texts can be copied.
 
 Fixed